I've done a lot of changes but not very significant.
First of all,
I've created newton_calcul and steepest_calcul. They don't compute the vectors usefull for the plot, so they are faster and shorter. I think we should put those ones in the report. They work exactky in the same way as newton and steepest but just shorter.
Secondly,
I've done calcul.m which is a function which calculate the data for the question i, I mean the execution time of each function and the residual for different values of tol and maxiterations. All the results are in the openoffice spreadsheet.